[mod] #115 위시리스트 페이지네이션 수정

+++ src/main/java/com/dmarket/dto/response/WishlistResDto.java
@@ -14,7 +14,7 @@ import com.dmarket.dto.common.WishlistItemDto;
 @AllArgsConstructor
 @Builder
 public class WishlistResDto {
-    private int wishCount;
+    private long wishCount;
     private List<WishlistItemDto> wishListItem;
     public int getWishCount() {
         return wishListItem.size();
